refactor: chart query keys
